refactor: TopBackground, TopBar 컴포넌트화
