+++ frontend/src/components/index.ts
@@ -1,2 +1,4 @@
 export { default as NavbarSection } from "./navbar/NavbarSection";
 export { default as ProfileBox } from "./ProfileBox";
+export { default as TopBackground } from "./common/TopBackground";
+export { default as TopBar } from "./common/TopBar";
